Refactor extension and inpage provider dependencies, remove unused code
- Updated `yarn.lock` to remove deprecated dependencies and add new ones for Solana and Keplr wallet support.
- Cleaned up `package.json` files by removing unused scripts and dependencies.
- Deleted obsolete utility files related to transaction handling and window management.
- Refactored transaction payload handling to streamline the integration of new transaction types and improve code clarity.
- Enhanced the Solana provider with improved transaction authority retrieval and simulation logic.
