+++ clients/extension/package.json
@@ -21,7 +21,6 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@bufbuild/protobuf": "^2.2.4",
     "@core/chain": "workspace:*",
     "@core/extension": "workspace:*",
     "@core/inpage-provider": "workspace:*",
@@ -35,7 +34,6 @@
     "@wallet-standard/base": "^1.1.0",
     "@wallet-standard/features": "^1.1.0",
     "bech32": "^2.0.0",
-    "bitcoinjs-lib": "^6.1.7",
     "mipd": "^0.0.7",
     "styled-components": "^6.1.16",
     "ua-parser-js": "^2.0.3"
