+++ clients/extension/src/navigation/AppView.ts
@@ -1,9 +1,5 @@
 import { CoreView } from '@core/ui/navigation/CoreView'
 
-export type AppView =
-  | CoreView
-  | { id: 'onboarding' }
-  | { id: 'connectedDapps' }
-  | { id: 'transactionTab' }
+export type AppView = CoreView | { id: 'onboarding' } | { id: 'connectedDapps' }
 
 export type AppViewId = AppView['id']
