--- clients/extension/src/types/thorchain.ts
@@ -1,1 +1,0 @@
-export type ThorchainProviderResponse = string | string[]
