+++ clients/extension/src/utils/functions.ts
@@ -1,57 +1,7 @@
-import { attempt, withFallback } from '@lib/utils/attempt'
-import { queryUrl } from '@lib/utils/query/queryUrl'
 import { VersionedTransaction } from '@solana/web3.js'
-
-export const calculateWindowPosition = (
-  currentWindow: chrome.windows.Window
-) => {
-  const height = 639
-  const width = 496
-  let left = 0
-  let top = 0
-
-  if (
-    currentWindow &&
-    currentWindow.left !== undefined &&
-    currentWindow.top !== undefined &&
-    currentWindow.width !== undefined
-  ) {
-    left = currentWindow.left + currentWindow.width - width
-    top = currentWindow.top
-  }
-
-  return { height, left, top, width }
-}
-
-export const checkERC20Function = async (
-  inputHex: string
-): Promise<boolean> => {
-  if (!inputHex || inputHex === '0x')
-    return new Promise(resolve => resolve(false))
-
-  const functionSelector = inputHex.slice(0, 10) // "0x" + 8 hex chars
-
-  const url = `https://www.4byte.directory/api/v1/signatures/?format=json&hex_signature=${functionSelector}&ordering=created_at`
-  const { count } = await withFallback(
-    attempt(() => queryUrl<{ count: number }>(url)),
-    { count: 0 }
-  )
-
-  return count > 0
-}
 
 export const isPopupView = () => {
   return chrome?.extension?.getViews({ type: 'popup' }).length > 0
-}
-
-export const splitString = (str: string, size: number): string[] => {
-  const result: string[] = []
-
-  for (let i = 0; i < str.length; i += size) {
-    result.push(str.slice(i, i + size))
-  }
-
-  return result
 }
 
 export function isVersionedTransaction(tx: any): tx is VersionedTransaction {
