+++ core/inpage-provider/package.json
@@ -9,8 +9,12 @@
     "@core/extension": "workspace:*",
     "@core/mpc": "workspace:*",
     "@core/ui": "workspace:*",
+    "@keplr-wallet/types": "^0.12.202",
     "@lib/extension": "workspace:*",
     "@lib/ui": "workspace:*",
-    "@lib/utils": "workspace:*"
+    "@lib/utils": "workspace:*",
+    "@solana/spl-token": "^0.4.13",
+    "@solana/web3.js": "^1.98.0",
+    "bitcoinjs-lib": "^6.1.7"
   }
 }
