+++ core/inpage-provider/tx/temp/core/solana/simulate/index.ts
