--- lib/extension/serialization/index.ts
@@ -1,1 +1,0 @@
-export * from '@core/inpage-provider/serialization'
